billing/src/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi-cli==0.0.4
fastapi-limiter==0.1.6
fastapi==0.111.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
loguru==0.7.2
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
pyasn1==0.6.0
pycparser==2.22
pydantic-core==2.18.2
pydantic-settings==2.2.1
pydantic==2.7.1
pygments==2.18.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pyyaml==6.0.1
redis==5.0.4
rich==13.7.1
rsa==4.9
sentry-sdk[loguru]==2.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
aiohttp==3.9.5
psycopg2==2.9.9
backoff==2.2.1
